move test data to mocks subdir
